Pass VMR arrays to platon

+++ src/proteus/config/_observe.py
@@ -12,14 +12,14 @@ class Platon:
 
     Attributes
     ----------
-    downsample: float
+    downsample: int
         Downsample binning factor for the spectrum.
     clip_vmr: float
         Minimum VMR for a species to be included in the radiative transfer.
     """
 
-    downsample: float   = field(default=8.0, validator=ge(1))
-    clip_vmr: float     = field(default=1e-8, validator=(gt(0), lt(1)))
+    downsample: int   = field(default=8, validator=ge(1))
+    clip_vmr: float   = field(default=1e-8, validator=(gt(0), lt(1)))
 
 @define
 class Observe:
